requirements.txt
# SEACABAr - Sistema Experto de Análisis Comparativo de Armas Balísticas
# Archivo unificado de dependencias
# Versión: 2.0.0

# ============================================================================
# DEPENDENCIAS PRINCIPALES (REQUERIDAS)
# ============================================================================

# Procesamiento de imágenes y visión por computadora
opencv-python==4.8.0.74
numpy==1.24.3
scipy==1.11.3
scikit-image==0.20.0
Pillow==10.0.1
rawpy==0.18.1
tifffile==2023.7.10

# Interfaz gráfica de usuario
PyQt5==5.15.9
PyQt5-tools==5.15.9.3.3

# Base de datos y búsqueda vectorial
faiss-cpu==1.7.4

# Machine Learning y análisis estadístico
scikit-learn==1.3.0
pandas==2.0.3

# Logging y configuración
loguru==0.7.0
pyyaml==6.0.1

# ============================================================================
# DEPENDENCIAS DE DESARROLLO Y TESTING
# ============================================================================

# Testing
pytest==7.4.0
pytest-qt==4.2.0
pytest-cov==4.1.0

# Empaquetado y distribución
pyinstaller==5.13.0
setuptools>=60.0.0
wheel>=0.37.0

# ============================================================================
# DEPENDENCIAS OPCIONALES
# ============================================================================

# APIs y servicios web (opcional - solo si se requiere funcionalidad web)
Flask==2.3.3
gunicorn==21.2.0
flask-cors==4.0.0

# GPU acceleration dependencies
opencv-contrib-python>=4.5.0  # OpenCV with CUDA support
cupy-cuda11x>=9.0.0  # CuPy for NumPy GPU acceleration (CUDA 11.x)
# cupy-cuda12x>=12.0.0  # Alternative for CUDA 12.x

# Deep Learning (opcional - solo para funcionalidades avanzadas)
# torch==2.0.1
# torchvision==0.15.2
# tensorflow==2.13.0

# Herramientas de desarrollo (opcional)
auto-py-to-exe==2.40.0

# ============================================================================
# NOTAS DE COMPATIBILIDAD
# ============================================================================
# - Python 3.8+ requerido
# - sqlite3, json, datetime, typing, dataclasses incluidos en Python estándar
# - Para Windows: instalar Visual C++ Redistributable
# - Para Linux: instalar libgl1-mesa-glx para OpenCV
# - Para macOS: instalar Xcode command line tools